+++ libmscore/shape.h
@@ -71,9 +71,9 @@ class Shape : public std::vector<ShapeElement> {
       qreal top() const;
       qreal bottom() const;
 
-      int size() const   { return std::vector<ShapeElement>::size(); }
-      bool empty() const { return std::vector<ShapeElement>::empty(); }
-      void clear()       { std::vector<ShapeElement>::clear();       }
+      size_t size() const { return std::vector<ShapeElement>::size();  }
+      bool empty() const  { return std::vector<ShapeElement>::empty(); }
+      void clear()        { std::vector<ShapeElement>::clear();        }
 
       bool contains(const QPointF&) const;
       bool intersects(const QRectF& rr) const;
